backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.0
django-filter==24.2
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
boto3==1.34.78
python-dotenv==1.0.1
